Add files via upload
This is an initial version of code to obtain an approximate solution of a differential equation. It requires the user to upload a trial solution for the given differential equation which consists of polynomial or sinusoidal terms, laden with coefficients to be determined based on boundary conditions and actual requirements of actually satisfying the differential equation.

It follows the theory mentioned in Chapter 2 of the book, "Textbook of Finite Element Analysis" by Prof. P. Seshu.

Example 2.6 from the textbook has been chosen as a particular case to test the code on. Subsequent improvements to the code will involve extension of this code to take into account further complexities in the trial solution and/or differential equation.

This is intended as a preamble to the objective of developing open source finite element code for structural analysis.